requirements.txt
google-cloud-pubsub==2.1.0
google-cloud-storage==1.32.0
google-cloud-translate==3.0.1
google-cloud-vision==2.0.0